mv MakeEmptyHeader to builder
