+++ execution/protocol/rules/misc/gaslimit.go
@@ -42,3 +42,20 @@ func VerifyGaslimit(parentGasLimit, headerGasLimit uint64) error {
 	}
 	return nil
 }
+
+// CalcGasLimit computes the gas limit of the next block after parent. It aims
+// to keep the baseline gas close to the provided target, and increase it towards
+// the target if the baseline gas is lower.
+func CalcGasLimit(parentGasLimit, desiredLimit uint64) uint64 {
+	delta := parentGasLimit/params.GasLimitBoundDivisor - 1
+	limit := parentGasLimit
+	desiredLimit = max(desiredLimit, params.MinBlockGasLimit)
+	// If we're outside our allowed gas range, we try to hone towards them
+	if limit < desiredLimit {
+		return min(parentGasLimit+delta, desiredLimit)
+	}
+	if limit > desiredLimit {
+		return max(parentGasLimit-delta, desiredLimit)
+	}
+	return limit
+}
